Make the Memoized metaclass use __hash__ instead of a homemade function
The use of  __hash__ make the whole memoization logic contained inside the
object itself, instead of relying on an external function (keyfunc).

The drawback is that an object can be created multiple times and destroyed
immediately after to leave only one instance, while with the past approach
no spurious creation is done.